feat(source-core): Added Core directory in ProjectLux module and moved the GameModeBase class to Core
- We want our core classes to be organized
