+++ Source/ProjectLux/Private/Core/ProjectLuxGameModeBase.cpp
@@ -1,5 +1,5 @@
 // Copyright Epic Games, Inc. All Rights Reserved.
 
 
-#include "ProjectLuxGameModeBase.h"
+#include "Core/ProjectLuxGameModeBase.h"
